Triplet::Repr is now inherited from Triplet. Minor refactoring
